chore: refactor statistic test folder structure
