+++ src/modules/statistic-test/components/others/contingency-table.tsx
@@ -6,8 +6,8 @@ import { useSelectLeftRightButtons } from '@/components/visual/select';
 import type { PlotParams } from 'react-plotly.js';
 import { fromPairs, max } from 'lodash-es';
 import { map2D, mask2D, sort2D, zip2D } from '@/common/utils/iterable';
-import { BaseStatisticTestResultRendererProps } from '../types';
-import { ContingencyTableConfig } from '../configuration/contingency-table';
+import { BaseStatisticTestResultRendererProps } from '../../types';
+import { ContingencyTableConfig } from '../../configuration/contingency-table';
 import {
   PlotInlineConfiguration,
   StatisticTestEmptyPlotWarning,
