+++ src/modules/statistic-test/components/others/subdataset-cooccurrence.tsx
@@ -7,7 +7,7 @@ import { Select, Stack } from '@mantine/core';
 import { PlotInlineConfiguration } from '@/modules/visualization/components/configuration';
 import { useDescriptionBasedRenderOption } from '@/components/visual/select';
 import { map2D, mask2D, zip2D } from '@/common/utils/iterable';
-import { BaseStatisticTestResultRendererProps } from '../types';
+import { BaseStatisticTestResultRendererProps } from '../../types';
 import { sum } from 'lodash-es';
 
 enum SubdatasetCoocccurrenceDisplayMode {
