+++ src/modules/statistic-test/components/plot-config.tsx
@@ -1,10 +1,8 @@
-import { useDescriptionBasedRenderOption } from '@/components/visual/select';
 import {
   Alert,
   Group,
   Input,
   type InputWrapperProps,
-  Select,
   Slider,
   Text,
 } from '@mantine/core';
@@ -106,56 +104,6 @@ export function useVisualizationMinFrequencySlider(
   );
 
   return { Component, threshold: debouncedMinFrequency, filter };
-}
-
-export enum BinaryStatisticTestVisualizationType {
-  Frequencies = 'frequencies',
-  ConfidenceLevel = 'significance',
-  EffectSize = 'effect-sizes',
-}
-
-const VISUALIZATION_TYPE_DICTIONARY = {
-  [BinaryStatisticTestVisualizationType.Frequencies]: {
-    label: 'Frequencies',
-    value: BinaryStatisticTestVisualizationType.Frequencies,
-    description:
-      'Show the frequencies of the rows that contains the categories.',
-  },
-  [BinaryStatisticTestVisualizationType.ConfidenceLevel]: {
-    label: 'Confidence Levels',
-    value: BinaryStatisticTestVisualizationType.ConfidenceLevel,
-    description:
-      'Show the confidence levels of the statistic tests, wherein the category/discriminator is used to split the dataset into two subdatasets that are compared against each other.',
-  },
-  [BinaryStatisticTestVisualizationType.EffectSize]: {
-    label: 'Effect Sizes',
-    value: BinaryStatisticTestVisualizationType.EffectSize,
-    description:
-      'Show the effect sizes of the statistic tests, wherein the category/discriminator is used to split the dataset into two subdatasets that are compared against each other.',
-  },
-};
-
-export function useBinaryStatisticTestVisualizationMethodSelect() {
-  const [type, setType] = React.useState(
-    BinaryStatisticTestVisualizationType.EffectSize,
-  );
-
-  const renderOption = useDescriptionBasedRenderOption(
-    VISUALIZATION_TYPE_DICTIONARY,
-  );
-
-  const Component = (
-    <Select
-      value={type}
-      onChange={setType as any}
-      data={Object.values(VISUALIZATION_TYPE_DICTIONARY)}
-      label="Data to Visualize"
-      renderOption={renderOption}
-      allowDeselect={false}
-    />
-  );
-
-  return { Component, type };
 }
 
 interface EmptyPlotWarningProps {
