+++ src/modules/statistic-test/components/regression/ordinal.tsx
@@ -26,7 +26,7 @@ import { generateColorsFromSequence } from '@/common/utils/colors';
 import { RegressionConfigType } from '../../configuration/regression-common';
 import { BaseStatisticTestResultRendererProps } from '../../types';
 import { ResultCard } from '@/components/visual/result-card';
-import { StatisticTestWarningsRenderer } from '../common';
+import { StatisticTestWarningsRenderer } from '../statistic-test/common';
 import { zip } from 'lodash-es';
 import { formatConfidenceLevel } from './utils';
 import { ToggleVisibility } from '@/components/visual/toggle-visibility';
