+++ src/modules/statistic-test/components/statistic-test/common.tsx
@@ -21,6 +21,7 @@ import {
   SimpleGrid,
   Switch,
   Spoiler,
+  Select,
 } from '@mantine/core';
 import { Warning } from '@phosphor-icons/react';
 import { max, sum } from 'lodash-es';
@@ -29,9 +30,11 @@ import {
   OMNIBUS_STATISTIC_TEST_METHOD_DICTIONARY,
   EFFECT_SIZE_DICTIONARY,
   OMNIBUS_EFFECT_SIZE_DICTIONARY,
-} from '../dictionary';
+} from '../../dictionary';
 import { generateColorsFromSequence } from '@/common/utils/colors';
 import { useDisclosure } from '@mantine/hooks';
+import React from 'react';
+import { useDescriptionBasedRenderOption } from '@/components/visual/select';
 
 interface StatisticTestWarningsRendererProps {
   warnings: string[] | null | undefined;
@@ -233,4 +236,54 @@ export function SignificanceAndEffectSizeComponents(
       <EffectSizeResultRenderer {...props.effectSize} />
     </SimpleGrid>
   );
+}
+
+export enum BinaryStatisticTestVisualizationType {
+  Frequencies = 'frequencies',
+  ConfidenceLevel = 'significance',
+  EffectSize = 'effect-sizes',
+}
+
+const VISUALIZATION_TYPE_DICTIONARY = {
+  [BinaryStatisticTestVisualizationType.Frequencies]: {
+    label: 'Frequencies',
+    value: BinaryStatisticTestVisualizationType.Frequencies,
+    description:
+      'Show the frequencies of the rows that contains the categories.',
+  },
+  [BinaryStatisticTestVisualizationType.ConfidenceLevel]: {
+    label: 'Confidence Levels',
+    value: BinaryStatisticTestVisualizationType.ConfidenceLevel,
+    description:
+      'Show the confidence levels of the statistic tests, wherein the category/discriminator is used to split the dataset into two subdatasets that are compared against each other.',
+  },
+  [BinaryStatisticTestVisualizationType.EffectSize]: {
+    label: 'Effect Sizes',
+    value: BinaryStatisticTestVisualizationType.EffectSize,
+    description:
+      'Show the effect sizes of the statistic tests, wherein the category/discriminator is used to split the dataset into two subdatasets that are compared against each other.',
+  },
+};
+
+export function useBinaryStatisticTestVisualizationMethodSelect() {
+  const [type, setType] = React.useState(
+    BinaryStatisticTestVisualizationType.EffectSize,
+  );
+
+  const renderOption = useDescriptionBasedRenderOption(
+    VISUALIZATION_TYPE_DICTIONARY,
+  );
+
+  const Component = (
+    <Select
+      value={type}
+      onChange={setType as any}
+      data={Object.values(VISUALIZATION_TYPE_DICTIONARY)}
+      label="Data to Visualize"
+      renderOption={renderOption}
+      allowDeselect={false}
+    />
+  );
+
+  return { Component, type };
 }
