+++ src/modules/statistic-test/configuration/contingency-table.tsx
@@ -2,29 +2,13 @@ import * as Yup from 'yup';
 
 import React from 'react';
 
-import { CATEGORICAL_SCHEMA_COLUMN_TYPES } from '@/api/project';
 import {
   ExcludeOverlappingRowsCheckbox,
+  StatisticTestMethodFormBody,
   StatisticTestOverlappingSubdatasetsWarning,
   StatisticTestProjectColumnSelectField,
 } from './utils';
 import { Stack } from '@mantine/core';
-
-export const binaryContingencyTableFormSchema = Yup.object({
-  column: Yup.string().required().default(''),
-});
-
-export type BinaryContingencyTableConfig = Yup.InferType<
-  typeof contingencyTableFormSchema
->;
-
-export function BinaryContingencyTableConfigForm() {
-  return (
-    <StatisticTestProjectColumnSelectField
-      supportedTypes={CATEGORICAL_SCHEMA_COLUMN_TYPES}
-    />
-  );
-}
 
 export const contingencyTableFormSchema = Yup.object({
   column: Yup.string().required().default(''),
@@ -39,7 +23,10 @@ export function ContingencyTableConfigForm() {
   return (
     <Stack>
       <StatisticTestOverlappingSubdatasetsWarning />
-      <BinaryContingencyTableConfigForm />
+      <StatisticTestProjectColumnSelectField
+        resets={['statistic_test_preference', 'effect_size_preference']}
+      />
+      <StatisticTestMethodFormBody />
       <ExcludeOverlappingRowsCheckbox />
     </Stack>
   );
